+++ front/app/containers/Admin/projects/project/analysis/InputsList/FieldValue.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { IInputsData } from 'api/analysis_inputs/types';
 import useIdeaCustomField from 'api/idea_custom_fields/useIdeaCustomField';
 
-import ShortFieldValue from '../components/ShortInputFieldValue';
+import ShortInputFieldValue from '../components/ShortInputFieldValue';
 
 type Props = {
   customFieldId: string;
@@ -17,12 +17,7 @@ type Props = {
  * representation of the value of the custom field for that input. Only renders
  * anything for non-built-in custom fields
  */
-const InputShortFieldValue = ({
-  projectId,
-  phaseId,
-  customFieldId,
-  input,
-}: Props) => {
+const FieldValue = ({ projectId, phaseId, customFieldId, input }: Props) => {
   const containerId: { projectId?: string; phaseId?: string } = {};
   if (projectId) {
     containerId.projectId = projectId;
@@ -39,7 +34,7 @@ const InputShortFieldValue = ({
   const rawValue =
     input.attributes.custom_field_values[customField.data.attributes.key];
 
-  return <ShortFieldValue customField={customField} rawValue={rawValue} />;
+  return <ShortInputFieldValue customField={customField} rawValue={rawValue} />;
 };
 
-export default InputShortFieldValue;
+export default FieldValue;
